# panoptikon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "panoptikon",
    "description": "",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/open/file/{sha256}": {
      "post": {
        "tags": [
          "open"
        ],
        "summary": "Open a file in the default application",
        "description": "Open a file in the default application on the host system.\nThis is done using os.startfile on Windows and xdg-open on Linux.\nThis is a potentially dangerous operation, as it can execute arbitrary code.",
        "operationId": "open_file_on_host",
        "parameters": [
          {
            "name": "index_db",
            "in": "query",
            "description": "The name of the `index` database to open and use for this API call. Find available databases with `/api/db`",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "user_data_db",
            "in": "query",
            "description": "The name of the `user_data` database to open and use for this API call. Find available databases with `/api/db`",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sha256",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File open request issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/open/folder/{sha256}": {
      "post": {
        "tags": [
          "open"
        ],
        "summary": "Show a file in the host system's file manager",
        "description": "Show a file in the host system's file manager.\nThis is done using the appropriate command for the host system.\nOn Windows, the file is highlighted in the Windows Explorer.\nOn macOS, the file is revealed in the Finder.\nThis is a potentially dangerous operation.",
        "operationId": "show_in_file_manager",
        "parameters": [
          {
            "name": "index_db",
            "in": "query",
            "description": "The name of the `index` database to open and use for this API call. Find available databases with `/api/db`",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "user_data_db",
            "in": "query",
            "description": "The name of the `user_data` database to open and use for this API call. Find available databases with `/api/db`",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sha256",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File explorer request issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OpenResponse": {
        "type": "object",
        "required": [
          "path",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}